Tests, more precise propagation
