+++ reactor-core/src/main/java/reactor/core/publisher/ContextPropagation.java
@@ -166,7 +166,6 @@ final class ContextPropagation {
 	}
 
 	static <T> SignalListener<T> contextRestoreForTap(final SignalListener<T> original, Supplier<Context> contextSupplier) {
-//		if (propagateContextToThreadLocals || !ContextPropagation.isContextPropagationAvailable()) {
 		if (!ContextPropagation.isContextPropagationAvailable()) {
 			return original;
 		}
