+++ reactor-core/src/main/java/reactor/core/publisher/MonoTapRestoringThreadLocals.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 VMware Inc. or its affiliates, All Rights Reserved.
+ * Copyright (c) 2022-2023 VMware Inc. or its affiliates, All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,13 +17,9 @@
 package reactor.core.publisher;
 
 import io.micrometer.context.ContextSnapshot;
-import org.reactivestreams.Subscription;
 import reactor.core.CoreSubscriber;
-import reactor.core.Exceptions;
-import reactor.core.Fuseable;
 import reactor.core.observability.SignalListener;
 import reactor.core.observability.SignalListenerFactory;
-import reactor.core.publisher.FluxTap.TapSubscriber;
 import reactor.util.annotation.Nullable;
 import reactor.util.context.Context;
 
@@ -80,7 +76,7 @@ final class MonoTapRestoringThreadLocals<T, STATE> extends MonoOperator<T, T> {
 		}
 
 		try (ContextSnapshot.Scope ignored = ContextPropagation.setThreadLocals(alteredContext)) {
-			source.subscribe(new TapSubscriber<>(actual, signalListener, alteredContext));
+			source.subscribe(new FluxTapRestoringThreadLocals.TapSubscriber<>(actual, signalListener, alteredContext));
 		}
 	}
 
@@ -91,240 +87,5 @@ final class MonoTapRestoringThreadLocals<T, STATE> extends MonoOperator<T, T> {
 		if (key == Attr.RUN_STYLE) return Attr.RunStyle.SYNC;
 
 		return super.scanUnsafe(key);
-	}
-
-	static class TapSubscriber<T> implements InnerOperator<T, T> {
-
-		final CoreSubscriber<? super T> actual;
-		final Context                   context;
-		final SignalListener<T>         listener;
-
-		boolean done;
-		Subscription s;
-
-		TapSubscriber(CoreSubscriber<? super T> actual, SignalListener<T> signalListener, Context ctx) {
-			this.actual = actual;
-			this.listener = signalListener;
-			this.context = ctx;
-		}
-
-		@Override
-		public CoreSubscriber<? super T> actual() {
-			return this.actual;
-		}
-
-		@Override
-		public Context currentContext() {
-			return this.context;
-		}
-
-		@Override
-		@Nullable
-		public Object scanUnsafe(Attr key) {
-			if (key == Attr.PARENT) return s;
-			if (key == Attr.TERMINATED) return done;
-			if (key == Attr.RUN_STYLE) return Attr.RunStyle.SYNC;
-
-			return InnerOperator.super.scanUnsafe(key);
-		}
-
-		/**
-		 * Cancel the prepared subscription, pass the listener error to {@link SignalListener#handleListenerError(Throwable)}
-		 * and then terminate the downstream directly with same error (without invoking any other handler).
-		 *
-		 * @param listenerError the exception thrown from a handler method before the subscription was set
-		 * @param toCancel the {@link Subscription} that was prepared but not sent downstream
-		 */
-		protected void handleListenerErrorPreSubscription(Throwable listenerError, Subscription toCancel) {
-			toCancel.cancel();
-			listener.handleListenerError(listenerError);
-			Operators.error(actual, listenerError);
-		}
-
-		/**
-		 * Cancel the active subscription, pass the listener error to {@link SignalListener#handleListenerError(Throwable)}
-		 * and then terminate the downstream directly with same error (without invoking any other handler).
-		 *
-		 * @param listenerError the exception thrown from a handler method
-		 */
-		protected void handleListenerErrorAndTerminate(Throwable listenerError) {
-			s.cancel();
-			listener.handleListenerError(listenerError);
-			actual.onError(listenerError); //TODO wrap ? hooks ?
-		}
-
-		/**
-		 * Cancel the active subscription, pass the listener error to {@link SignalListener#handleListenerError(Throwable)},
-		 * combine it with the original error and then terminate the downstream directly this combined exception
-		 * (without invoking any other handler).
-		 *
-		 * @param listenerError the exception thrown from a handler method
-		 * @param originalError the exception that was about to occur when handler was invoked
-		 */
-		protected void handleListenerErrorMultipleAndTerminate(Throwable listenerError, Throwable originalError) {
-			s.cancel();
-			listener.handleListenerError(listenerError);
-			RuntimeException multiple = Exceptions.multiple(listenerError, originalError);
-			actual.onError(multiple); //TODO wrap ? hooks ?
-		}
-
-		/**
-		 * After the downstream is considered terminated (or cancelled), pass the listener error to
-		 * {@link SignalListener#handleListenerError(Throwable)} then drop that error.
-		 *
-		 * @param listenerError the exception thrown from a handler method happening after sequence termination
-		 */
-		protected void handleListenerErrorPostTermination(Throwable listenerError) {
-			listener.handleListenerError(listenerError);
-			Operators.onErrorDropped(listenerError, actual.currentContext());
-		}
-
-		@Override
-		public void onSubscribe(Subscription s) {
-			try (ContextSnapshot.Scope ignored =
-						 ContextPropagation.setThreadLocals(actual.currentContext())) {
-				if (Operators.validate(this.s, s)) {
-					this.s = s;
-
-					try {
-						listener.doOnSubscription();
-					} catch (Throwable observerError) {
-						handleListenerErrorPreSubscription(observerError, s);
-						return;
-					}
-					actual.onSubscribe(this);
-				}
-			}
-		}
-
-		@Override
-		public void onNext(T t) {
-			try (ContextSnapshot.Scope ignored =
-						 ContextPropagation.setThreadLocals(actual.currentContext())) {
-				if (done) {
-					try {
-						listener.doOnMalformedOnNext(t);
-					} catch (Throwable observerError) {
-						handleListenerErrorPostTermination(observerError);
-					} finally {
-						Operators.onNextDropped(t, currentContext());
-					}
-					return;
-				}
-				try {
-					listener.doOnNext(t);
-				} catch (Throwable observerError) {
-					handleListenerErrorAndTerminate(observerError);
-					return;
-				}
-				actual.onNext(t);
-			}
-		}
-
-		@Override
-		public void onError(Throwable t) {
-			try (ContextSnapshot.Scope ignored =
-						 ContextPropagation.setThreadLocals(actual.currentContext())) {
-				if (done) {
-					try {
-						listener.doOnMalformedOnError(t);
-					} catch (Throwable observerError) {
-						handleListenerErrorPostTermination(observerError);
-					} finally {
-						Operators.onErrorDropped(t, currentContext());
-					}
-					return;
-				}
-				done = true;
-
-				try {
-					listener.doOnError(t);
-				} catch (Throwable observerError) {
-					//any error in the hooks interrupts other hooks, including doFinally
-					handleListenerErrorMultipleAndTerminate(observerError, t);
-					return;
-				}
-
-				actual.onError(t); //RS: onError MUST terminate normally and not throw
-
-				try {
-					listener.doAfterError(t);
-					listener.doFinally(SignalType.ON_ERROR);
-				} catch (Throwable observerError) {
-					handleListenerErrorPostTermination(observerError);
-				}
-			}
-		}
-
-		@Override
-		public void onComplete() {
-			try (ContextSnapshot.Scope ignored =
-						 ContextPropagation.setThreadLocals(actual.currentContext())) {
-				if (done) {
-					try {
-						listener.doOnMalformedOnComplete();
-					} catch (Throwable observerError) {
-						handleListenerErrorPostTermination(observerError);
-					}
-					return;
-				}
-				done = true;
-
-				try {
-					listener.doOnComplete();
-				} catch (Throwable observerError) {
-					handleListenerErrorAndTerminate(observerError);
-					return;
-				}
-
-				actual.onComplete(); //RS: onComplete MUST terminate normally and not throw
-
-				try {
-					listener.doAfterComplete();
-					listener.doFinally(SignalType.ON_COMPLETE);
-				} catch (Throwable observerError) {
-					handleListenerErrorPostTermination(observerError);
-				}
-			}
-		}
-
-		@Override
-		public void request(long n) {
-			try (ContextSnapshot.Scope ignored =
-						 ContextPropagation.setThreadLocals(this.context)) {
-				if (Operators.validate(n)) {
-					try {
-						listener.doOnRequest(n);
-					} catch (Throwable observerError) {
-						handleListenerErrorAndTerminate(observerError);
-						return;
-					}
-					s.request(n);
-				}
-			}
-		}
-
-		@Override
-		public void cancel() {
-			try (ContextSnapshot.Scope ignored =
-						 ContextPropagation.setThreadLocals(this.context)) {
-				try {
-					listener.doOnCancel();
-				} catch (Throwable observerError) {
-					handleListenerErrorAndTerminate(observerError);
-					return;
-				}
-
-				try {
-					s.cancel();
-				} finally {
-					try {
-						listener.doFinally(SignalType.CANCEL);
-					} catch (Throwable observerError) {
-						handleListenerErrorAndTerminate(observerError); //redundant s.cancel
-					}
-				}
-			}
-		}
 	}
 }
